# src/docs/swagger/openapi.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "SellMate API",
    "version": "1.0.0",
    "description": "\n# SellMate Platform - Social Commerce Automation Backend API\n\nSellMate is an intelligent social commerce automation platform designed specifically for Nigerian micro-entrepreneurs who sell through Instagram and WhatsApp.\n\n## Features\n- **Unified Conversations**: Consolidate WhatsApp and Instagram messages in one place\n- **AI-Powered Order Detection**: Automatically detect orders from conversations\n- **Profit Analytics**: Real-time business metrics and profit calculations\n- **Customer Management**: Track customer interactions across platforms\n\n## Authentication\nMost endpoints require JWT authentication. Include the token in the Authorization header:\n```\nAuthorization: Bearer <your_jwt_token>\n```\n\n## Rate Limiting\nAPI requests are rate-limited to prevent abuse. Current limits:\n- General: 100 requests per 15 minutes\n- Authentication: 5 requests per 15 minutes\n- AI endpoints: 20 requests per minute\n      ",
    "contact": {
      "name": "SellMate Team",
      "email": "support@sellmate.ng"
    },
    "license": {
      "name": "MIT",
      "url": "https://opensource.org/licenses/MIT"
    }
  },
  "servers": [
    {
      "url": "http://localhost:3000",
      "description": "Development server"
    },
    {
      "url": "https://api.sellmate.ng",
      "description": "Production server"
    }
  ],
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "Enter your JWT access token obtained from /auth/verify-otp or /auth/refresh"
      }
    },
...
  },
...
  "security": [
    {
      "bearerAuth": []
    }
  ],
  "paths": {
...
    "/api/v1/analytics/dashboard": {
      "get": {
        "summary": "Get dashboard metrics",
        "tags": [
          "Analytics"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "query",
            "name": "startDate",
            "required": true,
            "schema": {
              "type": "string",
              "format": "date-time"
            },
            "description": "Start date for analytics period"
          },
          {
            "in": "query",
            "name": "endDate",
            "required": true,
            "schema": {
              "type": "string",
              "format": "date-time"
            },
            "description": "End date for analytics period"
          }
        ],
        "responses": {
          "200": {
            "description": "Dashboard metrics including revenue, profit, orders, and customers",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean"
                    },
                    "data": {
                      "type": "object",
                      "properties": {
                        "revenue": {
                          "type": "object"
                        },
                        "profit": {
                          "type": "object"
                        },
                        "orders": {
                          "type": "object"
                        },
                        "customers": {
                          "type": "object"
                        }
                      }
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/analytics/revenue": {
      "get": {
        "summary": "Get revenue analytics",
        "tags": [
          "Analytics"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "query",
            "name": "startDate",
            "required": true,
            "schema": {
              "type": "string",
              "format": "date-time"
            }
          },
          {
            "in": "query",
            "name": "endDate",
            "required": true,
            "schema": {
              "type": "string",
              "format": "date-time"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Revenue analytics with platform breakdown and trends",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean"
                    },
                    "data": {
                      "type": "object",
                      "properties": {
                        "total": {
                          "type": "number"
                        },
                        "byPlatform": {
                          "type": "object",
                          "properties": {
                            "whatsapp": {
                              "type": "number"
                            },
                            "instagram": {
                              "type": "number"
                            },
                            "manual": {
                              "type": "number"
                            }
                          }
                        },
                        "trend": {
                          "type": "string",
                          "enum": [
                            "increasing",
                            "decreasing",
                            "stable"
                          ]
                        },
                        "percentageChange": {
                          "type": "number"
                        }
                      }
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/analytics/customers": {
      "get": {
        "summary": "Get customer analytics",
        "tags": [
          "Analytics"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "query",
            "name": "startDate",
            "required": true,
            "schema": {
              "type": "string",
              "format": "date-time"
            }
          },
          {
            "in": "query",
            "name": "endDate",
            "required": true,
            "schema": {
              "type": "string",
              "format": "date-time"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Customer analytics with segmentation and top customers",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean"
                    },
                    "data": {
                      "type": "object",
                      "properties": {
                        "totalCustomers": {
                          "type": "number"
                        },
                        "newCustomers": {
                          "type": "number"
                        },
                        "returningCustomers": {
                          "type": "number"
                        },
                        "topCustomers": {
                          "type": "array"
                        }
                      }
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/analytics/export": {
      "get": {
        "summary": "Export analytics report",
        "tags": [
          "Analytics"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "query",
            "name": "startDate",
            "required": true,
            "schema": {
              "type": "string",
              "format": "date-time"
            }
          },
          {
            "in": "query",
            "name": "endDate",
            "required": true,
            "schema": {
              "type": "string",
              "format": "date-time"
            }
          },
          {
            "in": "query",
            "name": "format",
            "schema": {
              "type": "string",
              "enum": [
                "json",
                "csv"
              ],
              "default": "json"
            }
          },
          {
            "in": "query",
            "name": "includeOrders",
            "schema": {
              "type": "boolean",
              "default": true
            }
          },
          {
            "in": "query",
            "name": "includeCustomers",
            "schema": {
              "type": "boolean",
              "default": true
            }
          },
          {
            "in": "query",
            "name": "includeProfit",
            "schema": {
              "type": "boolean",
              "default": true
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Analytics report file",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              },
              "text/csv": {
                "schema": {
                  "type": "string"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/customers": {
      "get": {
        "summary": "List customers for authenticated user",
        "tags": [
          "Customers"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "query",
            "name": "platform",
            "schema": {
              "type": "string",
              "enum": [
                "whatsapp",
                "instagram"
              ]
            },
            "description": "Filter by platform"
          },
          {
            "in": "query",
            "name": "search",
            "schema": {
              "type": "string"
            },
            "description": "Search by name or phone"
          },
          {
            "in": "query",
            "name": "page",
            "schema": {
              "type": "integer",
              "default": 1
            }
          },
          {
            "in": "query",
            "name": "limit",
            "schema": {
              "type": "integer",
              "default": 20
            }
          }
        ],
        "responses": {
          "200": {
            "description": "List of customers"
          },
          "401": {
            "description": "Unauthorized"
          }
        }
      },
      "post": {
        "summary": "Create a new customer",
        "tags": [
          "Customers"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "name"
                ],
                "properties": {
                  "name": {
                    "type": "string"
                  },
                  "phoneNumber": {
                    "type": "string"
                  },
                  "platforms": {
                    "type": "object",
                    "properties": {
                      "whatsapp": {
                        "type": "object",
                        "properties": {
                          "phoneNumber": {
                            "type": "string"
                          },
                          "profileName": {
                            "type": "string"
                          }
                        }
                      },
                      "instagram": {
                        "type": "object",
                        "properties": {
                          "username": {
                            "type": "string"
                          },
                          "profileName": {
                            "type": "string"
                          }
                        }
                      }
                    }
                  },
                  "notes": {
                    "type": "string"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Customer created"
          },
          "400": {
            "description": "Validation error"
          }
        }
      }
    },
...
  }
}